Fix int overflow in global stats total moji count

+++ Jiten.Api/Dtos/GlobalStatsDto.cs
@@ -6,5 +6,5 @@ public class GlobalStatsDto
 {
     public Dictionary<MediaType, int> MediaByType { get; set; } = new();
     public int TotalMedia { get; set; }
-    public float TotalMojis { get; set; }
+    public double TotalMojis { get; set; }
 }
